+++ package.json
@@ -8,7 +8,7 @@
     "cra-template": "1.2.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.29.5",
+    "react-router-dom": "^6.4.0",
     "react-scripts": "5.0.1"
   },
   "scripts": {
