--- src/styles/workplace-filter.css
@@ -1,39 +1,0 @@
-/* 검색 및 필터 컨테이너 */
-.workplace-filter {
-    display: flex;
-    gap: 10px;
-    margin-bottom: 20px;
-    align-items: center;
-  }
-  
-  /* 필터 입력 필드 */
-  .workplace-filter-input {
-    padding: 8px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    width: 150px;
-  }
-  
-  /* 필터 선택 드롭다운 */
-  .workplace-filter-select {
-    padding: 8px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-    background-color: white;
-    cursor: pointer;
-  }
-  
-  /* 검색 버튼 */
-  .workplace-filter-btn {
-    background-color: #007bff;
-    color: white;
-    border: none;
-    padding: 8px 15px;
-    border-radius: 5px;
-    cursor: pointer;
-  }
-  
-  .workplace-filter-btn:hover {
-    background-color: #0056b3;
-  }
-
